Save received messages in local db

+++ app/src/main/java/com/garlic/websockettest/MessageOverviewActivity.java
@@ -52,7 +52,7 @@ public class MessageOverviewActivity extends AppCompatActivity {
     }
 
     private void deleteMessages(){
-        double amount = msgHandler.resetMessages();
+        int amount = msgHandler.resetMessages();
         this.toast("Deleted messages: "+amount);
     }
